Add Husky w/o verification
