+++ .husky/pre-commit
@@ -1,0 +1,4 @@
+npm run lint:fix
+npm run format
+npm run check
+npm run test
