+++ app/(frontend)/new-poll/revalidate.tsx
@@ -2,5 +2,6 @@ import { revalidatePath } from "next/cache";
 
 export async function revalidateUserPolls() {
   "use server";
+
   revalidatePath("/user/polls");
 }
