+++ app/api/polls/[slug]/results/route.ts
@@ -1,4 +1,3 @@
-import { getData } from "@/app/(frontend)/polls/[slug]/getData";
 import { getPollResults } from "@/lib/pollResults/getPollResults";
 import { NextResponse } from "next/server";
