+++ app/components/polls/responses/Results.tsx
@@ -1,25 +1,20 @@
 "use client";
 
 import type { FC } from "react";
-import { useCallback, useMemo } from "react";
+import { useCallback } from "react";
 import type { SortKey, StatementWithStats } from "@/lib/pollResults/constants";
 import { CaretDownIcon } from "@radix-ui/react-icons";
-import type { Response, StatementOption } from "@/db/schema";
-import { useQuery } from "react-query";
-import { notFound, useParams, useSearchParams } from "next/navigation";
+import type { StatementOption } from "@/db/schema";
+import { useSearchParams } from "next/navigation";
 import type { getPollResults } from "@/lib/pollResults/getPollResults";
 import type { getData } from "@/app/(frontend)/polls/[slug]/getData";
-import { Statistics } from "./Statistics";
-import { ToggleGroup, ToggleGroupItem } from "../../shadcn/ui/toggle-group";
 import {
   Popover,
   PopoverTrigger,
   PopoverContent,
 } from "../../shadcn/ui/popover";
-
-type StatementWithStatsAndResponses = StatementWithStats & {
-  responses: Response[];
-};
+import { ToggleGroup, ToggleGroupItem } from "../../shadcn/ui/toggle-group";
+import { Statistics } from "./Statistics";
 
 export type ResultsProps = {
   initialResultsData: Awaited<ReturnType<typeof getPollResults>>;
@@ -108,150 +103,3 @@ export const Results: FC<ResultsProps> = ({
     />
   );
 };
-
-type EnabledDemographicFilters = Record<number, number[]>;
-
-export const useDemographicResponses = (
-  statements: StatementWithStatsAndResponses[],
-  statementOptions: Record<number, StatementOption[]>,
-  enabledDemographicFilters: Record<number, number[]>,
-) => {
-  // Pull out all demographic statements and responses
-
-  const demographicStatements = useMemo(
-    () =>
-      statements.filter(({ question_type }) => question_type === "demographic"),
-    [statements],
-  );
-
-  const allDemographicResponses = useMemo(
-    () => demographicStatements.flatMap((statement) => statement.responses),
-    [demographicStatements],
-  );
-
-  // Group responses of demographic questions by session_id
-
-  const demographicResponsesBySessionId = useMemo(
-    () =>
-      allDemographicResponses.reduce(
-        (acc, response) => {
-          if (!acc[response.session_id]) {
-            acc[response.session_id] = [];
-          }
-          acc[response.session_id].push(response);
-          return acc;
-        },
-        {} as Record<string, Response[]>,
-      ),
-    [allDemographicResponses],
-  );
-
-  // Let's get a list of sessionIds for each demographic option
-
-  const sessionIdsByDemographicOptionId = useMemo(
-    () =>
-      demographicStatements.reduce(
-        (acc, statement) =>
-          statementOptions[statement.id].reduce((optionCounts, option) => {
-            const sessionIds = allDemographicResponses
-              .filter((response) => response.option_id === option.id)
-              .map((response) => response.session_id);
-
-            return {
-              ...optionCounts,
-              [option.id]: sessionIds,
-            };
-          }, acc),
-        {} as Record<number, string[]>,
-      ),
-    [allDemographicResponses, demographicStatements, statementOptions],
-  );
-
-  // Get a count of how many sessions there are in each demographic group.
-  //
-  // For instance:
-  // If the user selects Age = <18, then we should update the counts such that the
-  // counts for other demographics are only for sessions that have Age = <18 (and
-  // set the count for Age = >=18 to 0).
-  //
-  // If no filters are set, then we should return the total number of sessions for
-  // each demographic group.
-  //
-  // Map this to a record of demographic option id -> count
-
-  const totalSessionCountsByDemographicOptionId = useMemo(
-    () =>
-      demographicStatements.reduce(
-        (acc, statement) =>
-          statementOptions[statement.id].reduce((countsAcc, option) => {
-            const sessionIds = sessionIdsByDemographicOptionId[option.id] ?? 0;
-
-            return {
-              ...countsAcc,
-              [option.id]:
-                Object.keys(enabledDemographicFilters).length > 0
-                  ? sessionIds.filter((sessionId) =>
-                      demographicResponsesBySessionId[sessionId].every(
-                        (response) =>
-                          response.option_id &&
-                          (enabledDemographicFilters[response.statementId]
-                            ?.length ?? 0) > 0
-                            ? enabledDemographicFilters[
-                                response.statementId
-                              ]?.includes(response.option_id)
-                            : true,
-                      ),
-                    ).length
-                  : sessionIds.length,
-            };
-          }, acc),
-        {} as Record<number, number>,
-      ),
-    [
-      demographicResponsesBySessionId,
-      demographicStatements,
-      enabledDemographicFilters,
-      sessionIdsByDemographicOptionId,
-      statementOptions,
-    ],
-  );
-
-  const response = useMemo(
-    () => ({
-      demographicStatements,
-      demographicResponsesBySessionId,
-      sessionIdsByDemographicOptionId,
-      totalSessionCountsByDemographicOptionId,
-    }),
-    [
-      demographicStatements,
-      demographicResponsesBySessionId,
-      sessionIdsByDemographicOptionId,
-      totalSessionCountsByDemographicOptionId,
-    ],
-  );
-
-  return response;
-};
-
-export function usePolledResultsData(
-  initialData: Awaited<ReturnType<typeof getPollResults>>,
-) {
-  const { slug } = useParams<{ slug: string }>();
-  const { data } = useQuery({
-    queryKey: ["/polls/[slug]/results", slug],
-    queryFn: async () => {
-      const res = await fetch(`/api/polls/${slug}/results`);
-      return res.json() as ReturnType<typeof getPollResults>;
-    },
-    initialData,
-    refetchInterval: 15_000,
-    staleTime: 15_000,
-  });
-
-  if (!data) {
-    notFound();
-  }
-
-  return data;
-}
